[OoT] Applied new organisation to ``c_writer/`` (#176)
* c_writer reorganised

* removed oot prefix

* black

* review

* moved scene/exporter content to a "to_c" subfolder
