--- fast64_internal/oot/c_writer/__init__.py
@@ -1,13 +1,0 @@
-from .oot_level_c import *
-from .oot_scene_table_c import *
-from .oot_spec import *
-from .oot_scene_folder import *
-from ..oot_level_parser import parseScene
-from .oot_scene_bootup import (
-    setBootupScene,
-    clearBootupScene,
-    ootSceneBootupRegister,
-    ootSceneBootupUnregister,
-    OOT_ClearBootupScene,
-    OOTBootupSceneOptions,
-)
