+++ fast64_internal/oot/scene/exporter/to_c/scene_collision.py
@@ -1,0 +1,9 @@
+from .....utility import CData
+from ....oot_collision import ootCollisionToC
+
+
+# Writes the collision data for a scene
+def ootSceneCollisionToC(scene):
+    sceneCollisionC = CData()
+    sceneCollisionC.append(ootCollisionToC(scene.collision))
+    return sceneCollisionC
